requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
click==7.1.2
Django==3.1.4
django-cors-headers==3.6.0
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
factory-boy==3.2.0
Faker==6.1.1
flake8==3.8.4
isort==5.7.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
